moved tests to /tests/litellm/litellm_core_utils
