+++ tests/litellm/litellm_core_utils/test_token_counter_tool_data.py
